+++ apps/researcher/src/app/[locale]/persons/person-card.tsx
@@ -2,7 +2,7 @@ import {useFormatter, useTranslations} from 'next-intl';
 import Link from 'next-intl/link';
 import {Person} from '@/lib/persons';
 import {PersonIcon} from '@/components/icons';
-import {encodeRouteSegment} from '@/lib/route-segment-transformer';
+import {encodeRouteSegment} from '@/lib/clerk-route-segment-transformer';
 
 interface Props {
   person: Person;
